добавляем/удаляем. Настраиваем логику работы слайдера.
