+++ js/main.js
@@ -6,7 +6,8 @@ import './generate-comments-template.js';
 import './big-picture-events.js';
 import './comments-functions.js';
 import './util.js';
-import './slider.js';
+import './slider/slider.js';
+import './slider/slider-functions.js';
 import './validation/validator.js';
 import './validation/hashtags-formatter.js';
 import './validation/validation-events.js';
