+++ js/slider/slider-functions.js
@@ -1,0 +1,9 @@
+import {uploadForm} from'../validation/validator.js';
+const uploadPreview = uploadForm.querySelector('.img-upload__preview');
+
+function test (value) {
+  uploadPreview.style.opacity = value;
+
+}
+
+export {test};
