+++ js/validation/validation-checks.js
@@ -12,7 +12,7 @@ const hashtagRules = [
   `Не более ${ MAX_HASHTAGS }ти хэштегов`,
   'Lлина комментария не может составлять больше 140 символов'
 ];
-
+// Обратите внимание, что при закрытии формы дополнительно необходимо сбрасывать значение поля выбора файла .img-upload__input. В принципе, всё будет работать, если при повторной попытке загрузить в поле другую фотографию. Но! Событие change не сработает, если пользователь попробует загрузить ту же фотографию, а значит окно с формой не отобразится, что будет нарушением техзадания. Значение других полей формы также нужно сбрасывать.
 
 const hashtagsRegular = /^#[a-zа-я0-9]{1,19}$/i;
